docs: documents createCmsImage
