+++ src/actions/cms/images/create.ts
@@ -1,23 +1,10 @@
 'use server'
-import prisma from '@/prisma'
-import { createPrismaActionError } from '@/actions/error'
 import type { Image } from '@prisma/client'
 import type { ExpandedCmsImage } from '@/cms/images/Types'
 import type { ActionReturn } from '@/actions/Types'
+import { createCmsImage } from '@/server/cms/images/create'
 
-export async function createCmsImage(name: string, image?: Image): Promise<ActionReturn<ExpandedCmsImage>> {
-    try {
-        const created = await prisma.cmsImage.create({
-            data: {
-                name,
-                imageId: image?.id
-            },
-            include: {
-                image: true
-            }
-        })
-        return { success: true, data: created }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+export async function createCmsImageAction(name: string, image?: Image): Promise<ActionReturn<ExpandedCmsImage>> {
+    //TODO: Auth route (very few people should be able to create stand alone cmsImages...)
+    return await createCmsImage(name, image)
 }
